# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "小红书爆款笔记生成 agent Skill",
    "description": "Enterprise-grade hot-note research, multi-model generation, account analytics and controlled publishing Agent Skill.",
    "version": "5.14.5"
  },
  "paths": {
...
    "/v1/accounts/{account_id}/publishing/drafts": {
      "post": {
        "tags": [
          "publishing"
        ],
        "summary": "Create Draft",
        "operationId": "create_draft_v1_accounts__account_id__publishing_drafts_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "account_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Account Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DraftRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublishDraft"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/publishing/drafts/{draft_id}/preview": {
      "post": {
        "tags": [
          "publishing"
        ],
        "summary": "Preview",
        "operationId": "preview_v1_publishing_drafts__draft_id__preview_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "draft_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Draft Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublishDraft"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/publishing/drafts/{draft_id}/preview-image": {
      "get": {
        "tags": [
          "publishing"
        ],
        "summary": "Preview Image",
        "operationId": "preview_image_v1_publishing_drafts__draft_id__preview_image_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "draft_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Draft Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/publishing/drafts/{draft_id}/approve": {
      "post": {
        "tags": [
          "publishing"
        ],
        "summary": "Approve",
        "operationId": "approve_v1_publishing_drafts__draft_id__approve_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "draft_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Draft Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApprovalRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublishApproval"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/publishing/drafts/{draft_id}/publish": {
      "post": {
        "tags": [
          "publishing"
        ],
        "summary": "Publish",
        "operationId": "publish_v1_publishing_drafts__draft_id__publish_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "draft_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Draft Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PublishRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublishResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/publishing/drafts/{draft_id}/schedule": {
      "post": {
        "tags": [
          "publishing"
        ],
        "summary": "Schedule",
        "operationId": "schedule_v1_publishing_drafts__draft_id__schedule_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "draft_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Draft Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScheduleRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublishSchedule"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/publishing/schedules/{schedule_id}": {
      "delete": {
        "tags": [
          "publishing"
        ],
        "summary": "Cancel Schedule",
        "operationId": "cancel_schedule_v1_publishing_schedules__schedule_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "schedule_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Schedule Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Cancel Schedule V1 Publishing Schedules  Schedule Id  Delete"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/providers": {
      "get": {
        "tags": [
          "providers"
        ],
        "summary": "Providers",
        "operationId": "providers_v1_providers_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Providers V1 Providers Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/v1/providers/{provider_id}/probe": {
      "post": {
        "tags": [
          "providers"
        ],
        "summary": "Probe",
        "operationId": "probe_v1_providers__provider_id__probe_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "provider_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Provider Id"
            }
          },
          {
            "name": "model",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Model"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Probe V1 Providers  Provider Id  Probe Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApprovalRequest": {
        "properties": {
          "ttl_minutes": {
            "type": "integer",
            "maximum": 120.0,
            "minimum": 1.0,
            "title": "Ttl Minutes",
            "default": 30
          },
          "ai_disclosure_confirmed": {
            "type": "boolean",
            "title": "Ai Disclosure Confirmed",
            "default": false
          },
          "commercial_disclosure_confirmed": {
            "type": "boolean",
            "title": "Commercial Disclosure Confirmed",
            "default": false
          },
          "account_identity_confirmed": {
            "type": "boolean",
            "title": "Account Identity Confirmed",
            "default": false
          },
          "enterprise_approval_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Enterprise Approval Id"
          }
        },
        "type": "object",
        "title": "ApprovalRequest"
      },
...
      "DeliveryPackage": {
        "properties": {
          "task_id": {
            "type": "string",
            "title": "Task Id"
          },
          "trace_id": {
            "type": "string",
            "title": "Trace Id"
          },
          "assumptions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Assumptions"
          },
          "research_summary": {
            "additionalProperties": true,
            "type": "object",
            "title": "Research Summary"
          },
          "hot_notes": {
            "items": {
              "$ref": "#/components/schemas/HotNoteCandidate"
            },
            "type": "array",
            "title": "Hot Notes"
          },
          "trend_insights": {
            "items": {
              "$ref": "#/components/schemas/TrendTopic"
            },
            "type": "array",
            "title": "Trend Insights"
          },
          "mechanisms": {
            "items": {
              "$ref": "#/components/schemas/ContentMechanism"
            },
            "type": "array",
            "title": "Mechanisms"
          },
          "strategy": {
            "additionalProperties": true,
            "type": "object",
            "title": "Strategy"
          },
          "title_candidates": {
            "items": {
              "$ref": "#/components/schemas/TitleCandidate"
            },
            "type": "array",
            "title": "Title Candidates"
          },
          "selected_title": {
            "type": "string",
            "title": "Selected Title"
          },
          "cover_options": {
            "items": {
              "$ref": "#/components/schemas/CoverOption"
            },
            "type": "array",
            "title": "Cover Options"
          },
          "body": {
            "type": "string",
            "title": "Body"
          },
          "graphic_pages": {
            "items": {
              "$ref": "#/components/schemas/GraphicPage"
            },
            "type": "array",
            "title": "Graphic Pages"
          },
          "video_script": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/VideoScript"
              },
              {
                "type": "null"
              }
            ]
          },
          "media_assets": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Media Assets"
          },
          "cover_asset": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cover Asset"
          },
          "topics": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Topics"
          },
          "location": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location"
          },
          "product_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Product Ids"
          },
          "keyword_map": {
            "additionalProperties": true,
            "type": "object",
            "title": "Keyword Map"
          },
          "hashtags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Hashtags"
          },
          "pinned_comment": {
            "type": "string",
            "title": "Pinned Comment",
            "default": ""
          },
          "cta": {
            "type": "string",
            "title": "Cta",
            "default": ""
          },
          "claims": {
            "items": {
              "$ref": "#/components/schemas/Claim"
            },
            "type": "array",
            "title": "Claims"
          },
          "originality_report": {
            "additionalProperties": true,
            "type": "object",
            "title": "Originality Report"
          },
          "compliance_report": {
            "additionalProperties": true,
            "type": "object",
            "title": "Compliance Report"
          },
          "ai_labeling": {
            "additionalProperties": true,
            "type": "object",
            "title": "Ai Labeling"
          },
          "quality_report": {
            "additionalProperties": true,
            "type": "object",
            "title": "Quality Report"
          },
          "content_hash": {
            "type": "string",
            "title": "Content Hash"
          },
          "publication_status": {
            "type": "string",
            "title": "Publication Status",
            "default": "HUMAN_REVIEW_REQUIRED"
          },
          "human_review_required": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Human Review Required"
          }
        },
        "type": "object",
        "required": [
          "task_id",
          "trace_id",
          "selected_title",
          "body",
          "content_hash"
        ],
        "title": "DeliveryPackage"
      },
...
      "DraftRequest": {
        "properties": {
          "package": {
            "$ref": "#/components/schemas/DeliveryPackage"
          },
          "mode": {
            "$ref": "#/components/schemas/PublishMode",
            "default": "REQUIRE_CONFIRMATION"
          }
        },
        "type": "object",
        "required": [
          "package"
        ],
        "title": "DraftRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PublishApproval": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "draft_id": {
            "type": "string",
            "title": "Draft Id"
          },
          "account_id": {
            "type": "string",
            "title": "Account Id"
          },
          "tenant_id": {
            "type": "string",
            "title": "Tenant Id",
            "default": "local"
          },
          "approved_by": {
            "type": "string",
            "title": "Approved By",
            "default": "local-cli"
          },
          "approver_auth_level": {
            "type": "integer",
            "title": "Approver Auth Level",
            "default": 1
          },
          "expected_content_hash": {
            "type": "string",
            "title": "Expected Content Hash"
          },
          "mode": {
            "$ref": "#/components/schemas/PublishMode",
            "default": "REQUIRE_CONFIRMATION"
          },
          "approval_token": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Approval Token"
          },
          "approval_token_hash": {
            "type": "string",
            "title": "Approval Token Hash"
          },
          "approved_at": {
            "type": "string",
            "format": "date-time",
            "title": "Approved At"
          },
          "expires_at": {
            "type": "string",
            "format": "date-time",
            "title": "Expires At"
          },
          "used_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Used At"
          },
          "scheduled_for": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scheduled For"
          },
          "ai_disclosure_confirmed": {
            "type": "boolean",
            "title": "Ai Disclosure Confirmed",
            "default": false
          },
          "commercial_disclosure_confirmed": {
            "type": "boolean",
            "title": "Commercial Disclosure Confirmed",
            "default": false
          },
          "account_identity_confirmed": {
            "type": "boolean",
            "title": "Account Identity Confirmed",
            "default": false
          },
          "enterprise_approval_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Enterprise Approval Id"
          }
        },
        "type": "object",
        "required": [
          "id",
          "draft_id",
          "account_id",
          "expected_content_hash",
          "approval_token_hash",
          "expires_at"
        ],
        "title": "PublishApproval"
      },
      "PublishDraft": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "account_id": {
            "type": "string",
            "title": "Account Id"
          },
          "tenant_id": {
            "type": "string",
            "title": "Tenant Id",
            "default": "local"
          },
          "created_by": {
            "type": "string",
            "title": "Created By",
            "default": "local-cli"
          },
          "package": {
            "$ref": "#/components/schemas/DeliveryPackage"
          },
          "content_hash": {
            "type": "string",
            "title": "Content Hash"
          },
          "mode": {
            "$ref": "#/components/schemas/PublishMode",
            "default": "REQUIRE_CONFIRMATION"
          },
          "preview_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Preview Url"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "account_id",
          "package",
          "content_hash"
        ],
        "title": "PublishDraft"
      },
      "PublishMode": {
        "type": "string",
        "enum": [
          "DRAFT_ONLY",
          "REQUIRE_CONFIRMATION",
          "SCHEDULED_AUTO_PUBLISH",
          "FULLY_AUTOMATED"
        ],
        "title": "PublishMode"
      },
      "PublishRequest": {
        "properties": {
          "approval_token": {
            "type": "string",
            "maxLength": 512,
            "minLength": 32,
            "title": "Approval Token"
          }
        },
        "type": "object",
        "required": [
          "approval_token"
        ],
        "title": "PublishRequest"
      },
      "PublishResult": {
        "properties": {
          "job_id": {
            "type": "string",
            "title": "Job Id"
          },
          "draft_id": {
            "type": "string",
            "title": "Draft Id"
          },
          "account_id": {
            "type": "string",
            "title": "Account Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "note_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note Url"
          },
          "note_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note Id"
          },
          "published_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Published At"
          },
          "failure_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Failure Code"
          },
          "failure_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Failure Message"
          },
          "audit": {
            "additionalProperties": true,
            "type": "object",
            "title": "Audit"
          }
        },
        "type": "object",
        "required": [
          "job_id",
          "draft_id",
          "account_id",
          "status"
        ],
        "title": "PublishResult"
      },
      "PublishSchedule": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "draft_id": {
            "type": "string",
            "title": "Draft Id"
          },
          "account_id": {
            "type": "string",
            "title": "Account Id"
          },
          "tenant_id": {
            "type": "string",
            "title": "Tenant Id",
            "default": "local"
          },
          "approval_id": {
            "type": "string",
            "title": "Approval Id"
          },
          "scheduled_at": {
            "type": "string",
            "format": "date-time",
            "title": "Scheduled At"
          },
          "status": {
            "type": "string",
            "title": "Status",
            "default": "SCHEDULED"
          },
          "failure_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Failure Message"
          }
        },
        "type": "object",
        "required": [
          "id",
          "draft_id",
          "account_id",
          "approval_id",
          "scheduled_at"
        ],
        "title": "PublishSchedule"
      },
...
      "ScheduleRequest": {
        "properties": {
          "approval_token": {
            "type": "string",
            "maxLength": 512,
            "minLength": 32,
            "title": "Approval Token"
          },
          "scheduled_at": {
            "type": "string",
            "format": "date-time",
            "title": "Scheduled At"
          }
        },
        "type": "object",
        "required": [
          "approval_token",
          "scheduled_at"
        ],
        "title": "ScheduleRequest"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}